# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Compute Platform API"
  version: "2024-01-15"
  description: >
    VM control plane API for instance lifecycle, images, volumes, snapshots,
    SSH keys, projects, and VPC networking.

    All mutating endpoints accept an optional `Idempotency-Key` header for
    retry-safe request deduplication.

    Authentication: `X-Principal-ID` header (Phase 1). Phase 2 targets
    SigV4 signed requests.

    All error responses follow the registered error envelope
    (see API_ERROR_CONTRACT_V1.md).

    ## Unsupported / Partially implemented

    - VPC/network endpoints: handler code is implemented and tested, but the
      routes are NOT yet wired into the production HTTP mux. Available only
      through the independent `NetworkHandlers` test seam.
    - Instance networking (NICs, SGs) at launch: gateway handler code exists
      but VPC network integration is incomplete.
    - Quota enforcement: count-based model active; reservation-based deferred.
    - RBAC: Phase 1 ownership model only (owner/principal). Phase 2 IAM not
      implemented.
    - KVM/real-machine runtime gate: all documented behaviour is proven at the
      handler seam via unit + integration tests. Hardware-gate validation
      requires Linux/KVM environment and is opt-in.
    - Image trust/validation pipeline: platform signature validation exists;
      user image import validation pipeline is minimal.
...
      (mTLS-protected) but not documented as a public endpoint — it is an
...
paths:
...
  /v1/instances:
    get:
      summary: "List instances"
      operationId: "listInstances"
      responses:
        "200":
          description: "Instance list (scoped to principal)"
      parameters:
        - name: "project_id"
          in: "query"
          schema:
            type: "string"
          description: "Optional: scope the list to a project"
    post:
      summary: "Create instance"
      operationId: "createInstance"
      responses:
        "202":
          description: "Instance creation accepted (async)"
        "400":
          description: "Invalid request"
        "401":
          description: "Missing or invalid authentication"
        "422":
          description: "Image not launchable, quota exceeded, or image family not found"
  /v1/instances/{id}:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
    get:
      summary: "Describe instance"
      operationId: "describeInstance"
      responses:
        "200":
          description: "Instance details"
        "404":
          description: "Not found or not owned"
    delete:
      summary: "Delete instance"
      operationId: "deleteInstance"
      responses:
        "202":
          description: "Delete accepted (async)"
  /v1/instances/{id}/stop:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
    post:
      summary: "Stop instance"
      operationId: "stopInstance"
      responses:
        "202":
          description: "Stop accepted"
        "409":
          description: "Illegal state transition"
...